+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- inventory: the ansible wrapper now able to produce inventories with
+per-host variables;
 - crypto: a method to export a single value secret as a file;
 - concord: make the organization name parameter optional for the
 inventory task;